+++ lizard_wbconfiguration/templates/lizard_wbconfiguration/wbconfiguration_read_only.js
@@ -1,0 +1,7 @@
+{% load history_extras %}
+{% load get_grid %}
+{% load get_portal_template %}
+{
+    xtype: 'wb_grid_history'
+
+}